+++ examples/hooks/index.js
@@ -1,27 +1,23 @@
 const feathers = require('@feathersjs/feathers');
 const mongoose = require('mongoose');
 const services = require('./services');
+const { Client } = require('../../lib');
 
 mongoose.Promise = global.Promise;
 mongoose.connect('mongodb://localhost:27017/feathers-mq', { useNewUrlParser: true });
 
 const app = feathers();
 
-app.set('appName', 'ServerProjectName');
+app.set('name', 'any_app');
 
 app.configure(services);
+app.configure(Client());
 
 (async function run() {
   await app.service('products').create({
-    name: 'anotherr',
+    name: 'another',
     price: 11,
   });
-
-  await app.service('products').find();
-
-  // app.service('categories').create({
-  //   name: 'beer',
-  // });
 }());
 
 console.log('running server instance');
