+++ examples/server/index.js
@@ -6,7 +6,7 @@ const { Server } = require('../../lib');
 
 mongoose.Promise = global.Promise;
 
-mongoose.connect('mongodb://localhost/feathers-mq');
+mongoose.connect('mongodb://localhost:27017/feathers-mq', { useNewUrlParser: true });
 
 const app = feathers();
 
@@ -15,4 +15,4 @@ app.set('name', 'ServerProjectName');
 app.configure(services);
 app.configure(Server());
 
-console.log('running server instance');
+console.log('running instance with appName:', app.get('name'));
