+++ lib/hooks/index.js
@@ -1,15 +1,14 @@
 const debug = require('debug')('feathers-mq:hook:');
-const NATS = require('nats');
 const Errors = require('@feathersjs/errors');
+const getNats = require('../instance');
 
 const deniedMethods = ['find', 'get'];
 
 module.exports = function Logger() {
-  
-  const nats = NATS.connect({ json: true });
+  const nats = getNats();
 
   return function logger(context) {
-    const appName = context.app.get('appName');
+    const appName = context.app.get('name');
 
     if (!appName) {
       throw new Errors.BadRequest('App name is required');
@@ -17,7 +16,7 @@ module.exports = function Logger() {
 
     const serviceName = context.path;
     const methodName = context.method;
-    
+
     if (deniedMethods.includes(context.method)) {
       debug('denied method ', methodName);
       return context;
@@ -30,6 +29,7 @@ module.exports = function Logger() {
       data: context.data,
       params: context.params,
       type: context.type,
+      app: appName,
     };
 
     const natsSubject = `${appName}.${serviceName}.${methodName}`;
