+++ lib/index.js
@@ -1,9 +1,9 @@
 const Server = require('./mq-server');
 const Client = require('./mq-client');
-const hooks = require('./hooks');
+const logger = require('./hooks');
 
 module.exports = {
   Server,
   Client,
-  hooks,
+  logger,
 };
