+++ lib/mq-client/index.js
@@ -2,17 +2,21 @@ const debug = require('debug')('feathers-mq:client:a');
 const NATS = require('nats');
 const Errors = require('@feathersjs/errors');
 
+let nats = null;
+
 module.exports = function getFn(options = {}) {
-  return function attachService() {
-    const app = this;
+  if (!nats) {
     const host = options.host || 'localhost';
     const port = options.ports || 4222;
+    nats = NATS.connect(`nats://${host}:${port}`);
+  }
+
+  return function attachService() {
+    const app = this;
 
     if (!app.get('name')) {
       throw new Errors.BadRequest('App name is required');
     }
-
-    const nats = NATS.connect(`nats://${host}:${port}`);
 
     nats.on('error', (err) => {
       // exit process
@@ -32,7 +36,7 @@ module.exports = function getFn(options = {}) {
       process.exit(1);
     });
 
-    debug(`Connected to NATS on ${host}:${port} and requesting from ${app.get('name')}.*`);
+    debug(`Connected to NATS as Client : ${app.get('name')}.*`);
 
     app.service = function natsService(name) {
       let finalName = name;
@@ -70,7 +74,6 @@ module.exports = function getFn(options = {}) {
 
       return {
         find(params = {}) {
-          // debug("Find was requested")
           const method = 'find';
           const request = {
             params,
@@ -125,14 +128,12 @@ module.exports = function getFn(options = {}) {
           return sendRequest(method, request);
         },
 
-        on(evt, cb) {
-          return nats.subscribe(`${name}.${evt}.events`, async (request) => {
-            await cb(JSON.parse(request));
-          });
-        },
+        // on(evt, cb) {
+        //   return nats.subscribe(`${name}.${evt}.events`, async (request) => {
+        //     await cb(JSON.parse(request));
+        //   });
+        // },
       };
     };
-
-    debug('feathers mq client ran');
   };
 };
